requirements.txt
# Airflow
apache-airflow-providers-databricks==6.0.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-dbt-cloud==3.7.0

# Database
psycopg2-binary==2.9.9

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0